+++ R/zzz.R
@@ -12,6 +12,7 @@
   register_s3_method("tidyr", "pivot_wider", "data.table")
   register_s3_method("tidyr", "replace_na", "data.table")
   register_s3_method("tidyr", "nest", "data.table")
+  register_s3_method("tidyr", "separate", "data.table")
 
   register_s3_method("dplyr", "filter", "dtplyr_step")
   register_s3_method("dplyr", "intersect", "dtplyr_step")
@@ -25,6 +26,7 @@
   register_s3_method("tidyr", "pivot_wider", "dtplyr_step")
   register_s3_method("tidyr", "replace_na", "dtplyr_step")
   register_s3_method("tidyr", "nest", "dtplyr_step")
+  register_s3_method("tidyr", "separate", "dtplyr_step")
 }
 
 register_s3_method <- function(pkg, generic, class, fun = NULL) {