+++ tests/testthat/test-tidyeval.R
@@ -102,6 +102,12 @@ test_that("translates case_when()", {
     quote(fcase(x1, y1, x2, y2, x3, TRUE, rep(TRUE, .N), y4))
   )
 
+  # can use T for default, #272
+  expect_equal(
+    capture_dot(dt, case_when(x1 ~ y1, x2 ~ y2, x3 ~ TRUE, T ~ y4)),
+    quote(fcase(x1, y1, x2, y2, x3, TRUE, rep(TRUE, .N), y4))
+  )
+
   # translates recursively
   expect_equal(
     capture_dot(dt, case_when(x == 1 ~ n())),